backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.2
scikit-learn==1.3.2
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1

